Applications/Streamlit/requirements.txt
streamlit==1.31.0
pandas==2.1.4
gitpython==3.1.42
openpyxl==3.1.2
requests==2.31.0 